serde: move version.h to envelope.h
