+++ src/v/serde/envelope.h
@@ -9,14 +9,24 @@
 
 #pragma once
 
-#include "serde/version.h"
-
 #include <cinttypes>
 #include <concepts>
 #include <cstddef>
 #include <type_traits>
 
 namespace serde {
+
+using version_t = std::uint8_t;
+
+template<version_t V>
+struct version {
+    static constexpr auto const v = V;
+};
+
+template<version_t V>
+struct compat_version {
+    static constexpr auto const v = V;
+};
 
 /**
  * \brief provides versioning (version + compat version)
