+++ src/v/serde/read_header.h
@@ -15,7 +15,6 @@
 #include "serde/rw/scalar.h"
 #include "serde/serde_size_t.h"
 #include "serde/type_str.h"
-#include "serde/version.h"
 #include "serde_exception.h"
 #include "ssx/sformat.h"
 #include "vlog.h"
